refactor API into class
